requirements.txt
azure-functions
httpx[http2]
aiolimiter
